+++ CHANGELOG.md
@@ -7,11 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+- Show error dialog and exit if GPU does not support OpenGL 4.4
+
 ### Fixed
 - Bug where changing multiple scattering probability did not trigger a new
   render if maximum iterations were reached
 - Crashing on Intel GPUs with Linux and Mesa3D
-
 
 ## 2.3.1 - 2019-07-14